# docs/swagger.yaml
definitions:
  be-dashboard-nba_internal_presentation_presenter.Pagination:
    properties:
      has_next:
        example: true
        type: boolean
      has_prev:
        example: false
        type: boolean
      page:
        example: 1
        type: integer
      page_size:
        example: 20
        type: integer
      total_items:
        example: 100
        type: integer
      total_pages:
        example: 5
        type: integer
    type: object
...
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.Pagination'
    type: object
...
  be-dashboard-nba_internal_presentation_presenter_user.CreateUserRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      full_name:
        type: string
      name:
        type: string
      password:
        minLength: 8
        type: string
      phone:
        type: string
      role_id:
        type: integer
    required:
    - active
    - email
    - full_name
    - name
    - password
    - role_id
    type: object
  be-dashboard-nba_internal_presentation_presenter_user.ReadUserDetailResponse:
    properties:
      active:
        type: boolean
      email:
        type: string
      full_name:
        type: string
      id:
        type: string
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      role_id:
        type: integer
    type: object
  be-dashboard-nba_internal_presentation_presenter_user.ReadUserResponse:
    properties:
      active:
        type: boolean
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  be-dashboard-nba_internal_presentation_presenter_user.UpdateUserRequest:
    properties:
      active:
        type: boolean
      email:
        type: string
      full_name:
        type: string
      name:
        type: string
      phone:
        type: string
      role_id:
        type: integer
    required:
    - active
    - email
    - full_name
    - name
    - role_id
    type: object
info:
  contact: {}
  description: Template API for Golang project
  title: Base APP Go
  version: "1.0"
paths:
...
  /users:
    get:
      description: Retrieves a paginated and searchable list of users in the system.
        Requires Bearer token.
      parameters:
      - description: Search term for name or email
        in: query
        name: search
        type: string
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 10)'
        in: query
        name: limit
        type: integer
      - description: Sort order (e.g., name ASC)
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get users
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadPaginate'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.ReadUserResponse'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Failed get users
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Creates a new user with the provided details. Requires Bearer token.
      parameters:
      - description: User data to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully create user
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid JSON or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Role ID not found
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      tags:
      - User
  /users/{user_id}:
    delete:
      description: Deletes a user by their ID. Requires Bearer token. Users cannot
        delete themselves.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully delete user
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid User ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "403":
          description: Forbidden (Attempt to delete own account)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (User ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      tags:
      - User
    get:
      description: Fetches the details of a single user by their ID. Requires Bearer
        token.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully get user detail
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
            - properties:
                data:
                  $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.ReadUserDetailResponse'
              type: object
        "400":
          description: Bad Request (Invalid User ID)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: Not Found (User ID not found)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates a user's profile details by their ID. Requires Bearer token.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update user
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: User ID not found
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      tags:
      - User
  /users/me:
    get:
      description: Fetches the profile details of the currently authenticated user.
        Requires Bearer token.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user profile
          schema:
            allOf:
            - $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
            - properties:
                data:
                  $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.ReadUserResponse'
              type: object
        "401":
          description: Unauthorized (Invalid or missing token)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "404":
          description: User profile not found
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Get Current User Profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates a user's profile details by their ID. Requires Bearer token.
      parameters:
      - description: User data to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter_user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully update user
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadData'
        "400":
          description: Bad Request (Invalid ID, Invalid JSON, or Validation Error)
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/be-dashboard-nba_internal_presentation_presenter.ResponsePayloadMessage'
      security:
      - BearerAuth: []
      summary: Update User
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {your token}" into the field below.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"